Motherboard Snap-On
This part can be used, to snap the motherboard brackets onto the X frame.
Print at 0.2, it's a tight fit. If you want to remove them from the X frame, push a screwdriver underneath the part where the motherboard bracket goes in. So you can push it out.
That's the only way I found to remove them from the frame :)